src/app/page/contact/contact.component.ts
import {Component} from '@angular/core';
import {FormBuilder, FormGroup, ReactiveFormsModule, Validators} from "@angular/forms";
import {ButtonComponent} from "../../component/button/button.component";

@Component({
  selector: 'app-contact',
  standalone: true,
  imports: [
    ReactiveFormsModule,
    ButtonComponent
  ],
  templateUrl: './contact.component.html',
})
export class ContactComponent {

  public contactForm: FormGroup;

  constructor(private formBuilder: FormBuilder) {
    this.contactForm = this.formBuilder.group({
      name: ['', Validators.required],
      email: ['', Validators.required],
      phone: ['', Validators.required],
      comment: ['', Validators.required]
    })
  }

  public submitForm() {
    console.log('Form submitted:', this.contactForm.value);
  }
}
